python-package/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "upnify-mcp-server"
version = "1.0.4"
description = "MCP Server for Upnify CRM integration - reports, prospects, opportunities, and reminders"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "Adriel Lisa", email = "tu-email@ejemplo.com" },
]
keywords = [
    "mcp",
    "model-context-protocol", 
    "upnify",
    "crm",
    "ai",
    "claude",
    "sales",
    "prospects"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial :: Accounting",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/adrielisa/MCP"
Repository = "https://github.com/adrielisa/MCP.git"
Issues = "https://github.com/adrielisa/MCP/issues"

[project.scripts]
upnify-mcp-server = "upnify_mcp_server.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/upnify_mcp_server"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/pyproject.toml",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
target-version = "py38"
